ParallaxView now stores disposables in its own dispose bag.
